requirements.txt
fastapi==0.115.12
httpx==0.28.1
redis>=4.5.0
uvicorn==0.33.0
python-dotenv==1.0.1